+++ cpp/ql/src/Likely Bugs/Arithmetic/SignedOverflowCheck-good2.cpp
@@ -1,3 +1,3 @@
-bool baz(int n1, unsigned int delta) {
-    return n1 + delta < n1; // GOOD
+bool baz(int n1, int delta) {
+    return (unsigned)n1 + delta < n1; // GOOD
 }
